Mixins example
